+++ package.json
@@ -35,6 +35,7 @@
 		"build": "rm -rf ./dist/ && webpack --mode=production --config config/webpack.config.js",
 		"build:example": "rm -rf ./examples/dist/ && webpack --mode=production --config=./examples/webpack.config.js",
 		"clean": "rm -rf ./dist/ ./examples/dist/",
+		"dev:types": "tsc --noEmit --watch",
 		"pre:publish": "npm run build && npm run build:example && npm run test",
 		"start": "rm -rf ./dist/ && webpack --mode=development --config config/webpack.config.js",
 		"start:example": "rm -rf ./examples/dist/ && webpack serve --mode=development --config=./examples/webpack.config.js",
