+++ src/components/control-bar/control-bar.css
