+++ src/components/control-bar/templates/control-bar.ts
