+++ src/components/overlay/overlay.ts
@@ -1,0 +1,9 @@
+import './overlay.css'
+
+/**
+ * Overlay template
+ * @returns {String} Generated HTML
+ */
+export default function overlay(): string {
+	return '<div class="v-overlay"></div>'
+}
