+++ src/components/poster/poster.css
