--- src/plugins/airplay/config.js
@@ -1,7 +1,0 @@
-// import CSS
-import './css/airplay.css'
-
-// import JS
-import airplay from './js/airplay'
-
-export default airplay
