--- src/plugins/cast/config.js
@@ -1,7 +1,0 @@
-// import CSS
-import './css/cast.css'
-
-// import JS
-import cast from './js/cast'
-
-export default cast
