--- src/plugins/ima/config.js
@@ -1,7 +1,0 @@
-// import CSS
-import './css/ima.css'
-
-// import JS
-import ima from './js/ima'
-
-export default ima
