--- src/plugins/pip/config.js
@@ -1,7 +1,0 @@
-// import CSS
-import './css/pip.css'
-
-// import JS
-import pictureInPicture from './js/pip'
-
-export default pictureInPicture
