--- src/plugins/subtitle/config.js
@@ -1,7 +1,0 @@
-// import CSS
-import './css/subtitle.css'
-
-// import JS
-import subtitle from './js/subtitle'
-
-export default subtitle
