--- src/plugins/volume-bar/config.js
@@ -1,7 +1,0 @@
-// import CSS
-import './css/volume-bar.css'
-
-// import JS
-import volumeBar from './js/volume-bar'
-
-export default volumeBar
