+++ src/plugins/volume-bar/volume-bar.css
