--- src/vlite/components/big-play/config.js
@@ -1,5 +1,0 @@
-// import CSS
-import './assets/styles/big-play.css'
-
-// import JS
-import './assets/scripts/big-play'
