--- src/vlite/components/control-bar/config.js
@@ -1,5 +1,0 @@
-// import CSS
-import './assets/styles/control-bar.css'
-
-// import JS
-import './assets/scripts/control-bar'
