--- src/vlite/components/loader/config.js
@@ -1,5 +1,0 @@
-// import CSS
-import './assets/styles/loader.css'
-
-// import JS
-import './assets/scripts/loader'
