--- src/vlite/components/overlay/assets/scripts/overlay.ts
@@ -1,7 +1,0 @@
-/**
- * Overlay template
- * @returns {String} Generated HTML
- */
-export default function overlay(): string {
-	return '<div class="v-overlay"></div>'
-}
