--- src/vlite/components/overlay/config.js
@@ -1,5 +1,0 @@
-// import CSS
-import './assets/styles/overlay.css'
-
-// import JS
-import './assets/scripts/overlay'
