--- src/vlite/components/poster/config.js
@@ -1,5 +1,0 @@
-// import CSS
-import './assets/styles/poster.css'
-
-// import JS
-import './assets/scripts/poster'
