--- src/vlite/config.js
@@ -1,13 +1,0 @@
-// Import JS
-import Vlitejs from './assets/scripts/vlite'
-
-import './assets/styles/vlite.css'
-
-// import shared components
-import './components/loader/config.js'
-import './components/control-bar/config.js'
-import './components/big-play/config.js'
-import './components/overlay/config.js'
-import './components/poster/config.js'
-
-export default Vlitejs
